Add new utility script
The purpose of this script is to generate a subset of the variant data (and corresponding metadata) that exclude contributions from one or more sources.  This works with either variants_output.tsv (with variants_output_field_metadata.tsv) or build_final.tsv (with field_metadata.tsv).  Given the variant file and one or more source to exclude, this script (1) removes from the variant data the columns that pertain to the excluded sources, (2) removes mention of the excluded sources from the Source column, and (3) removes any variants that are specific to the excluded sources.  It then generates a metadata file that omits the fields that pertain to the excluded sources.