Update config.beta and prod js files
config.beta.js and config.prod.js file format is out of date and does not match files as
actually deployed. Update files, omitting various keys.
(These keys are publicly visible in brcaexchange source so this is not for security reasons but
to prevent out-of-date or incorrect keys from being used accidentally.)